Can generate and place multiple snowflakes.

+++ snowflake.h
@@ -34,7 +34,7 @@ snowflake* initSnowflake(int, int, int, int);
 
 
 /*  Snowflake Operations  */
-void setOrigin(snowflake*, int, int, int);
+void setOrigin(snowflake*, double, double, double);
 void setEllipses(snowflake*, int, int, int);
 void combineGeom(snowflake*, snowflake*);
 int boxCollide(snowflake*, snowflake*);
@@ -42,9 +42,9 @@ void import2DArr(snowflake*, double*, int);
 void updateMaxMin(snowflake*);
 
 /*  Snowflake Translations  */
-void rotate(snowflake*, float, float, float, float);
-void translate(snowflake*, float, float, float);
-void scale(snowflake*, float, float ,float);
+void rotate(snowflake*, double, double, double, double);
+void translate(snowflake*, double, double, double);
+void scale(snowflake*, double, double ,double);
 
 
 /*rotate:       angle, x, y, z
